Standalone usage of /p2p-circuit addr (dependency gardening)
Removal of the old constant is going to be a separate step.
