+++ multiaddr/multiaddr.go
@@ -1,0 +1,18 @@
+package multiaddr
+
+import (
+	ma "github.com/multiformats/go-multiaddr"
+)
+
+const P_CIRCUIT = 290
+
+var Protocol = ma.Protocol{
+	Code:  P_CIRCUIT,
+	Size:  0,
+	Name:  "p2p-circuit",
+	VCode: ma.CodeToVarint(P_CIRCUIT),
+}
+
+func init() {
+	ma.AddProtocol(Protocol)
+}
